feat: update header to 26.10

+++ src/mc/world/actor/ActorResetRule.h
@@ -2,7 +2,7 @@
 
 #include "mc/_HeaderOutputPredefine.h"
 
-enum class ActorResetRule : int {
+enum class ActorResetRule : uchar {
     Default        = 0,
     LoadedFromSave = 1,
     Respawn        = 2,
